fix: OS nav bar is covering the pay modal
